requirements.txt

aiohttp
beautifulsoup4
feedparser
pandas
numpy
yfinance
ccxt
python-dotenv
pyyaml
transformers
torch
scikit-learn
chromadb
sentence-transformers
tqdm
faker
pytest
requests
datasets
accelerate
streamlit
plotly
watchdog
tweepy>=4.0.0
async-lru
oauthlib
